Only remove MLNX openmpi if present
Quick-Functional: true
Skip-coverity-test: true
Skip-func-test-vm: true
Skip-func-hw-test-small: false
Skip-func-hw-test-medium: true
Skip-func-hw-test-large: true
Skip-scan-centos-rpms: true
Skip-scan-leap15-rpms: true
Skip-test-centos-rpms: true
Allow-unstable-test: true
PR-repos-el8: gotestsum@PR-6
RPM-test-version: 1.3.105
Skip-PR-comments: true


#Pragmas from previous commit message:
Quick-Functional: true
Skip-coverity-test: true
Skip-func-test-vm: true
Skip-func-hw-test-small: false
Skip-func-hw-test-medium: true
Skip-func-hw-test-large: true
Skip-scan-centos-rpms: true
Skip-scan-leap15-rpms: true
Skip-test-centos-rpms: true
Test-tag: test_always_passes
Allow-unstable-test: true
PR-repos-el8: gotestsum@PR-6
RPM-test-version: 1.3.105
Skip-PR-comments: true
Quick-Functional: true
Skip-coverity-test: true
Skip-func-test-vm: true
Skip-func-hw-test-small: false
Skip-func-hw-test-medium: true
Skip-func-hw-test-large: true
Skip-scan-centos-rpms: true
Skip-test-centos-rpms: true
Test-tag: test_always_passes
Allow-unstable-test: true
PR-repos-el8: gotestsum@PR-6
RPM-test-version: 1.3.105
Skip-PR-comments: true
Quick-Functional: true
Skip-coverity-test: true
Skip-func-test-vm: true
Skip-func-hw-test-small: false
Skip-func-hw-test-medium: true
Skip-func-hw-test-large: true
Skip-scan-centos-rpms: true
Skip-test-centos-rpms: true
Test-tag: test_always_passes
Allow-unstable-test: true
PR-repos-el8: gotestsum@PR-6
RPM-test-version: 1.3.105
Skip-PR-comments: true

+++ ci/provisioning/post_provision_config_nodes_EL_8.sh
@@ -87,9 +87,9 @@ post_provision_config_nodes() {
     fi
     if [ -n "$INST_RPMS" ]; then
         # Remove the mellanox packages first to allow the distro packages to be installed
-        ls -d /usr/mpi/gcc/openmpi-* || true
-        rpm -q openmpi opensm-libs
-        dnf -y erase openmpi opensm-libs
+        if ls -d /usr/mpi/gcc/openmpi-*; then
+            dnf -y erase openmpi opensm-libs
+        fi
         # shellcheck disable=SC2086
         time dnf -y erase $INST_RPMS
     fi
